fixed pulling main ui to my branch and updating everything
