+++ src/main/java/edu/wpi/teamc/controllers/FloorPlanController.java
@@ -1,5 +1,7 @@
 package edu.wpi.teamc.controllers;
 
+import edu.wpi.teamc.navigation.Navigation;
+import edu.wpi.teamc.navigation.Screen;
 import javafx.event.ActionEvent;
 import javafx.fxml.FXML;
 import javafx.scene.control.MenuButton;
@@ -38,4 +40,55 @@ public class FloorPlanController {
 
   @FXML
   void getMenuButton(ActionEvent event) {}
+
+  @FXML
+  void getFlowerDeliveryPage(ActionEvent event) {
+    Navigation.navigate(Screen.FLOWER);
+  }
+
+  @FXML
+  void getFurnitureDeliveryPage(ActionEvent event) {
+    Navigation.navigate(Screen.FURNITURE);
+  }
+
+  @FXML
+  void getHelpPage(ActionEvent event) {
+    Navigation.navigate(Screen.HELP);
+  }
+
+  @FXML
+  void getMealDeliveryPage(ActionEvent event) {
+    Navigation.navigate(Screen.MEAL);
+  }
+
+  @FXML
+  void getOfficeSuppliesPage(ActionEvent event) {
+    Navigation.navigate(Screen.OFFICE_SUPPLY);
+  }
+
+  @FXML
+  void getRoomReservationPage(ActionEvent event) {
+    Navigation.navigate(Screen.CONFERENCE);
+  }
+
+  @FXML
+  void getSignagePage(ActionEvent event) {
+    Navigation.navigate(Screen.SIGNAGE);
+  }
+
+  /** Method run when controller is initialized */
+  @FXML
+  public void initialize() {}
+
+  @FXML
+  void getEditMap(ActionEvent event) {}
+
+  @FXML
+  void getLogOut(ActionEvent event) {}
+
+  @FXML
+  void getMapHistory(ActionEvent event) {}
+
+  @FXML
+  void getMapPage(ActionEvent event) {}
 }
