+++ src/main/java/edu/wpi/teamc/controllers/PathfindingController.java
@@ -143,4 +143,51 @@ public class PathfindingController {
     }
     return false;
   }
+
+  @FXML
+  void getFlowerDeliveryPage(ActionEvent event) {
+    Navigation.navigate(Screen.FLOWER);
+  }
+
+  @FXML
+  void getFurnitureDeliveryPage(ActionEvent event) {
+    Navigation.navigate(Screen.FURNITURE);
+  }
+
+  @FXML
+  void getHelpPage(ActionEvent event) {
+    Navigation.navigate(Screen.HELP);
+  }
+
+  @FXML
+  void getMealDeliveryPage(ActionEvent event) {
+    Navigation.navigate(Screen.MEAL);
+  }
+
+  @FXML
+  void getOfficeSuppliesPage(ActionEvent event) {
+    Navigation.navigate(Screen.OFFICE_SUPPLY);
+  }
+
+  @FXML
+  void getRoomReservationPage(ActionEvent event) {
+    Navigation.navigate(Screen.CONFERENCE);
+  }
+
+  @FXML
+  void getSignagePage(ActionEvent event) {
+    Navigation.navigate(Screen.SIGNAGE);
+  }
+
+  @FXML
+  void getEditMap(ActionEvent event) {}
+
+  @FXML
+  void getLogOut(ActionEvent event) {}
+
+  @FXML
+  void getMapHistory(ActionEvent event) {}
+
+  @FXML
+  void getMapPage(ActionEvent event) {}
 }
